backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks and job queue
celery==5.3.4
flower==2.0.1

# Virtualization integrations
proxmoxer==2.0.1
paramiko==3.3.1
fabric==3.2.2
docker==6.1.3

# VPN management
cryptography==41.0.7
ipaddress

# File handling and storage
boto3==1.34.0
minio==7.2.0

# Monitoring and metrics
prometheus-client==0.19.0
influxdb-client==1.38.0

# Email
fastapi-mail==1.4.1

# WebSocket support
python-socketio==5.10.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
loguru==0.7.2

# Date and time handling
python-dateutil==2.8.2
